Added wiki#journal#make_index function
Added a small make_index function to journal.vim that collects all
journal files and inserts an index with links to all the journal
files into the current file at the location of the cursor. The index
is sorted with years as the top level set of headers, months as the
second level, and links to entries as a bulleted list beneath the month
headers.

This is very similar to the :VimwikiDiaryGenerateLinks command from
Vimwiki