Add fzf support
This adds `:WikiFzfPages` and `:WikiFzfToc` commands which allows to
jump between wiki pages and between toc entries within a single page
correspondingly. Also corresponding `<Plug>` mappings are added which
should work both in normal and in insert modes (usually fzf keybindings
are used with a `<C-*>` key).

I've called them `WikiFzf*` and not `FzfWiki*` because I'm thinking of
them of being wiki specific first. That's inconstent with the current
CtrlP integration though.