Update wiki.txt
On line 499(500) I've changed "... `pdf` files."  to  "... output files." to clarify that not only pdf are available (as hinted with 'test.html'). Here "... exported files." could also be an option. I don't know what would be best.

On line 687 I change `wiki_pdf_viewer` to `wiki_viwer` as this seems to be the correct one (not `wiki_file_open`, right?). `wiki_pdf_viewer` is not found  anywhere else in the doc.

Thanks for this awesome plugin!