Add support for Orgmode style tags.
This adds a new tags parser in addition to the default one:
g:wiki#tags#orgmode_parser. It's very similar to the default with the
following differences:

- *only* supports short-style tags `:tag1:tag2:`;
- orgmode only allows tags in two specific instances: at the end of
  headlines and in the `#+FILETAGS: :tag1:tag2:` property at the
  beginning of a file.

The parser uses these restrictions in the regexes that search for tags.
The only reason that we can't use the default parser is that its
`make()` method generates long style tags `:tag1: :tag2:`, which are
forbidden in orgmode.

The orgmode parser is added to the package, but not to any of the
default settings to avoid interference with the default parser. People
who wish to use it probably are best served explicitly opting in via
g:wiki_tag_parsers.