Add support for renaming of Orgmode files.
This is a rather straight-forward change. The file
autoload/wiki/page.vim is adjusted to also take orgmode links with
descriptions into account and a unit test for it is added. The function
s:convert_links_to_html() is also extended for orgmode links, though I'm
not sure how to test it.