Add support for Orgmode style links.
Orgmode links look like this: [[link]], [[link][description]]. This is
very similar, but not identical to wiki links.

This adds a new matcher to autoload/wiki/link/. It is carefully added to
s:matchers so that cycling between wiki, markdown and orgmode links
never causes a deadlock.

The relevant regexes are added to autoload/wiki/rx.vim and the
documentation is updated. This also includes the new matcher in the
config variable g:wiki_link_toggles. Some orgmode links are added to the
wiki-basic test wiki and the affected unit tests (such as counting
links) are updated. The test test/test-markdown/test.vim is updated such
as to toggle the existing link from Markdown to Orgmode styl The test
test/test-markdown/test.vim is updated such as to toggle the existing
link from markdown to orgmode style.