feat!: remove support for Unite and Denite
I don't want to keep support for these interfaces. I instead want to
provide an API that makes it easy to build such interfaces on top of for
those who are interested. It should be easy to copy the last versions
of these files to ones personal configuration to keep the
functionality.