feat: major refactoring of url handlers/resolvers
This changes the implementation of how URLs are resolved and handled.
The old style was somewhat convoluted in that the URLs were
unnecessarily entwined with the link concept. The new implementation
more properly separates the concepts. This is a major step towards
allowing more flexibility for users to define their own URL resolvers
and handlers.

In the process, the g:wiki_resolver option was removed. This will
temporarily _remove_ some flexibility. However, the idea is that the
new ideas should allow much more flexibility.

refer: #302