Allow complete customization of setting link target
When I use `WikiLinkAdd` and select a file, the link target is always an
absolute path from the wiki root, dispite I try to customize it in vimrc
file. What I would like to use is actually a relative path from the
inserted page to the target page. Both are in the same wiki root. The
existing behavior will always insert an absolute path.

In this change, the customized function will always be called to
transform the link target, if the customized function exists. Then it'll
fall back to the existing logic of using the absolute path.